analysis-server/requirements.txt
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
kafka-python==2.0.2
torch==2.1.0
torchvision==0.16.0
numpy==1.26.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1
protobuf>=3.20.0
grpcio>=1.41.0
geventhttpclient>=1.4.4
tritonclient==2.41.0
